Modify rad2deg

+++ src/ft_math.c
@@ -6,7 +6,7 @@
 /*   By: Yoshihiro Kosaka <ykosaka@student.42tok    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/09 13:03:00 by ykosaka           #+#    #+#             */
-/*   Updated: 2023/02/02 09:44:10 by Yoshihiro K      ###   ########.fr       */
+/*   Updated: 2023/02/06 22:07:41 by Yoshihiro K      ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,14 +17,14 @@ long long	ft_math_dtoll(double num)
 	return ((long long)num);
 }
 
-double	ft_math_deg2rad(int deg)
+double	ft_math_deg2rad(double deg)
 {
 	return (M_PI * deg / 180);
 }
 
 double	ft_math_rad2deg(double rad)
 {
-	return ((int)(rad * 180 / M_PI));
+	return (rad * 180 / M_PI);
 }
 
 double	ft_math_angle_normalize(double angle)
